MediaPlayer instance singleton with interruptions care + Gallery bugs fixed (pick and take)

+++ app/src/main/java/com/example/puzzlegame/model/MusicSettings.java
@@ -34,7 +34,7 @@ public class MusicSettings {
         this.playingMusic = play;
     }
 
-    public double getMusicVol() {
+    public int getMusicVol() {
         return musicVol;
     }
